+++ packages/manager/src/mocks/presets/extra/account/child-account.ts
@@ -1,12 +1,13 @@
 import { http } from 'msw';
 
-import { MockPreset } from 'src/mocks/mockPreset';
 import {
   accountFactory,
-  profileFactory,
   accountUserFactory,
+  profileFactory,
 } from 'src/factories';
 import { makeResponse } from 'src/mocks/utilities/response';
+
+import type { MockPreset } from 'src/mocks/types';
 
 /**
  * Mocks account, profile, and user requests to simulate a Parent/Child child account user.
@@ -23,16 +24,16 @@ const mockChildAccount = () => {
     http.get('*/v4*/profile', () => {
       return makeResponse(
         profileFactory.build({
-          username: 'Parent Account User',
           user_type: 'proxy',
+          username: 'Parent Account User',
         })
       );
     }),
     http.get(`*/v4*/account/users/Parent Account User`, () => {
       return makeResponse(
         accountUserFactory.build({
-          username: 'Parent Account User',
           user_type: 'proxy',
+          username: 'Parent Account User',
         })
       );
     }),
@@ -40,9 +41,9 @@ const mockChildAccount = () => {
 };
 
 export const childAccountPreset: MockPreset = {
-  label: 'Child Account Proxy User',
-  id: 'parent-child-child-account-proxy-user',
   desc: 'Mock a Parent/Child child account proxy user',
   group: 'Account',
   handlers: [mockChildAccount],
+  id: 'parent-child-child-account-proxy-user',
+  label: 'Child Account Proxy User',
 };
