+++ packages/manager/src/mocks/presets/extra/environment/regions.ts
@@ -1,10 +1,11 @@
 import { getRegions } from 'src/mocks/handlers/region-handlers';
-import type { MockPreset } from 'src/mocks/mockPreset';
+
+import type { MockPreset } from 'src/mocks/types';
 
 export const regionsPreset: MockPreset = {
-  label: 'Regions',
-  id: 'regions',
   desc: 'Mock Linode regions',
   group: 'Environment',
   handlers: [getRegions],
+  id: 'regions',
+  label: 'Regions',
 };
