+++ packages/manager/src/mocks/presets/extra/linodes/linodes-crud.ts
@@ -1,9 +1,10 @@
-import { getLinodes, createLinodes } from 'src/mocks/handlers/linode-handlers';
-import { MockPreset } from 'src/mocks/mockPreset';
+import { createLinodes, getLinodes } from 'src/mocks/handlers/linode-handlers';
+
+import type { MockPreset } from 'src/mocks/types';
 
 export const linodeCrudPreset: MockPreset = {
-  label: 'Linode CRUD',
-  id: 'linodes-crud',
   group: 'Linodes',
   handlers: [getLinodes, createLinodes],
+  id: 'linodes-crud',
+  label: 'Linode CRUD',
 };
